fixed abstract class issue with saving ID and refetching places instead, added viepager2, created placesrepository
issues with out of bounds for swiping and major issues with order of and allocation of images to the right restaurant in the right order.

routing through RestaurantsFragment.kt also seemed to completely remove the issue

+++ app/src/main/java/com/example/foodie/data/model/DiscoverySettings.kt
@@ -1,7 +1,7 @@
 package com.example.foodie.data.model
 
 data class DiscoverySettings(
-    var radius: Long = 2000,
+    var radius: Double = 2000.0,
     var openNow: Boolean = true,
     var minRating: Float = 0f,
     var priceLevels: MutableList<String> = mutableListOf(),
